backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
apispec==6.8.0
attrs==24.2.0
bcrypt==4.2.1
black==24.10.0
blinker==1.9.0
cfgv==3.4.0
claid==0.6.4
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dateutils==0.6.12
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
filelock==3.16.1
flake8==7.1.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
flask-pydantic-spec==0.6.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
fonttools==4.55.1
grpcio==1.59.3
gunicorn==23.0.0
identify==2.6.3
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
mccabe==0.7.0
mypy==1.13.0
mypy-extensions==1.0.0
narwhals==1.15.2
nodeenv==1.9.1
numpy==1.25.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
protobuf==4.25.3
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==1.10.19
pydantic_core==2.27.1
pyflakes==3.2.0
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
six==1.16.0
SQLAlchemy==2.0.36
tomli==2.2.1
types-Flask-Cors==5.0.0.20240902
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.28.0
vl-convert-python==1.7.0
webargs==8.6.0
Werkzeug==3.1.3
zipp==3.21.0
